+++ src/make_matrix.h
@@ -9,12 +9,12 @@
     struct NAME                                                                                    \
     {                                                                                              \
         T*            data;                                                                        \
-        uint_fast32_t nrows;                                                                       \
-        uint_fast32_t ncols;                                                                       \
+        uint_fast16_t nrows;                                                                       \
+        uint_fast16_t ncols;                                                                       \
     };
 
 #define MAKE_MATRIX_CREATE(NAME, T)                                                                \
-    static inline struct NAME* NAME##_create(uint_fast32_t const nrows, uint_fast32_t const ncols) \
+    static inline struct NAME* NAME##_create(uint_fast16_t const nrows, uint_fast16_t const ncols) \
     {                                                                                              \
         struct NAME* matrix = malloc(sizeof(struct NAME));                                         \
         matrix->data = malloc(sizeof(T) * (nrows * ncols));                                        \
@@ -24,26 +24,26 @@
     }
 
 #define MAKE_MATRIX_GET(NAME, T)                                                                   \
-    static inline T NAME##_get(struct NAME const* matrix, uint_fast32_t r, uint_fast32_t c)        \
+    static inline T NAME##_get(struct NAME const* matrix, uint_fast16_t r, uint_fast16_t c)        \
     {                                                                                              \
         return matrix->data[r * matrix->ncols + c];                                                \
     }
 
 #define MAKE_MATRIX_GET_PTR(NAME, T)                                                               \
-    static inline T* NAME##_get_ptr(struct NAME const* matrix, uint_fast32_t r, uint_fast32_t c)   \
+    static inline T* NAME##_get_ptr(struct NAME const* matrix, uint_fast16_t r, uint_fast16_t c)   \
     {                                                                                              \
         return matrix->data + (r * matrix->ncols + c);                                             \
     }
 
 #define MAKE_MATRIX_GET_PTR_C(NAME, T)                                                             \
-    static inline T const* NAME##_get_ptr_c(struct NAME const* matrix, uint_fast32_t r,            \
-                                            uint_fast32_t c)                                       \
+    static inline T const* NAME##_get_ptr_c(struct NAME const* matrix, uint_fast16_t r,            \
+                                            uint_fast16_t c)                                       \
     {                                                                                              \
         return matrix->data + (r * matrix->ncols + c);                                             \
     }
 
 #define MAKE_MATRIX_SET(NAME, T)                                                                   \
-    static inline void NAME##_set(struct NAME* matrix, uint_fast32_t r, uint_fast32_t c, T v)      \
+    static inline void NAME##_set(struct NAME* matrix, uint_fast16_t r, uint_fast16_t c, T v)      \
     {                                                                                              \
         matrix->data[r * matrix->ncols + c] = v;                                                   \
     }
@@ -51,7 +51,7 @@
 #define MAKE_MATRIX_SET_ALL(NAME, T)                                                               \
     static inline void NAME##_set_all(struct NAME* matrix, T v)                                    \
     {                                                                                              \
-        for (uint_fast32_t i = 0; i < matrix->nrows * matrix->ncols; ++i)                          \
+        for (uint_fast16_t i = 0; i < matrix->nrows * matrix->ncols; ++i)                          \
             matrix->data[i] = v;                                                                   \
     }
 
@@ -66,14 +66,14 @@
     }
 
 #define MAKE_MATRIX_NROWS(NAME)                                                                    \
-    static inline uint_fast32_t NAME##_nrows(struct NAME const* matrix) { return matrix->nrows; }
+    static inline uint_fast16_t NAME##_nrows(struct NAME const* matrix) { return matrix->nrows; }
 
 #define MAKE_MATRIX_NCOLS(NAME)                                                                    \
-    static inline uint_fast32_t NAME##_ncols(struct NAME const* matrix) { return matrix->ncols; }
+    static inline uint_fast16_t NAME##_ncols(struct NAME const* matrix) { return matrix->ncols; }
 
 #define MAKE_MATRIX_RESIZE(NAME, T)                                                                \
-    static inline void NAME##_resize(struct NAME* matrix, uint_fast32_t const nrows,               \
-                                     uint_fast32_t const ncols)                                    \
+    static inline void NAME##_resize(struct NAME* matrix, uint_fast16_t const nrows,               \
+                                     uint_fast16_t const ncols)                                    \
     {                                                                                              \
         matrix->data = realloc(matrix->data, sizeof(T) * (nrows * ncols));                         \
         matrix->nrows = nrows;                                                                     \
